requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
fpdf==1.7.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pillow==10.3.0
psycopg==3.1.19
psycopg2==2.9.9
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
setuptools==69.2.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
wheel==0.43.0
